requirements.txt
aniso8601==8.0.0
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.13.1
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-Dance==3.0.0
Flask-Login==0.5.0
flask-paginate==0.5.4
Flask-RESTful==0.3.8
Flask-Session==0.3.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
oauthlib==3.1.0
pycodestyle==2.6.0
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.12.0
SQLAlchemy==1.3.17
SQLAlchemy-Utils==0.36.5
urllib3==1.25.9
URLObject==2.4.3
Werkzeug==0.16.0
